requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
pandas>=1.5.0

# ML / Science
scikit-learn>=1.2.0
scikit-image>=0.20.0

# NLP / Transformers (VLM language branch)
transformers>=4.30.0
tokenizers>=0.13.0

# Config
pyyaml>=6.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress
tqdm>=4.65.0

# Logging / Tracking
tensorboard>=2.13.0

# Dataset download
kaggle>=1.5.16

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
